tweak

+++ app/lib/batch.py
@@ -11,7 +11,7 @@ class Batch(Base):
     with current_app.test_request_context():
       q = Queue("batch")
       maxSentenceLen = TrainingSet.maxSentenceLen()
-      training_set_count = TrainingSet.count()/batchSize
+      training_set_count = int(TrainingSet.count()/batchSize)
       for j in range(0, max_epoch):
         for i in range(0, training_set_count):
           training_sets = TrainingSet.order_by_raw("random()").paginate(batchSize, i)
